requirements.txt
Flask==2.2.3
flask_cors==3.0.10

python-dotenv
prettyprint

pandas

pytest

PyPDF2